adds UserContactUpdater to move updating contact and availability logic
