+++ app/models/concerns/user_info_updater.rb
@@ -1,12 +1,12 @@
 require 'active_support/concern'
 
 module UserInfoUpdater
+  include UserContactUpdater
   include ActiveSupport::Concern
 
   def self.update(user_id, user_type, user_params)
-    update_contact(user_id, user_params)
+    UserContactUpdater.update(user_id, user_params)
     update_interests(user_id, user_params[:interests]) if user_params[:interests]
-    update_availability(user_id, user_params[:availability]) if user_params[:availability]
     update_identities(user_id, user_params[:identities]) if user_params[:identities]
 
     if user_type == :mentor
@@ -22,27 +22,11 @@ module UserInfoUpdater
     UserNonTechSkill.update_for_user(user_id, skills.map(&:to_i)) if type == :non_tech
   end
 
-  def self.update_availability(user_id, availability_params)
-     new_availability = availability_params.to_h.reduce({}) do |new_avail, (day_of_week, availability)|
-       new_avail[day_of_week.to_i] = availability.map{|period| ActiveModel::Type::Boolean.new.cast(period)}
-       new_avail
-     end
-     Availability.update_for_user(user_id, new_availability)
-  end
-
   def self.update_identities(user_id, identities)
     UserIdentity.update_for_user(user_id, identities.map(&:to_i))
   end
 
   def self.update_interests(user_id, interests)
     UserInterest.update_for_user(user_id, interests.map(&:to_i))
-  end
-
-  def self.update_contact(user_id, user_params)
-    updated_attributes = {user_id: user_id}
-    [:email, :slack, :linkedin, :phone].each do |param|
-      updated_attributes[param] = user_params[param] if user_params[param]
-    end
-    ContactDetails.update(updated_attributes)
   end
 end
